cli/requirements.txt
argparse==1.4.0
python-crontab==2.3.5
tabulate==0.8.2
PyYAML==5.1.0
jsonschema==3.0.1
ansible-vault==1.2.0
joblib==0.13.2